+++ NEWS.md
@@ -1,5 +1,6 @@
 # taxonbridge 1.0.4
 
+* Updated `get_validity()` method
 * Fixed bug in `load_taxonomies()` when reading in GBIF data with vroom
 * Updated vignette
 * Added `annotate()` method